Move ServerFramework.{h,cpp} into server directory
