+++ msvc2017/FreeOrionD/FreeOrionD.vcxproj
@@ -144,11 +144,11 @@
     <ClInclude Include="..\..\network\Networking.h" />
     <ClInclude Include="..\..\network\ServerNetworking.h" />
     <ClInclude Include="..\..\python\CommonFramework.h" />
-    <ClInclude Include="..\..\python\server\ServerFramework.h" />
     <ClInclude Include="..\..\python\server\ServerWrapper.h" />
     <ClInclude Include="..\..\python\SetWrapper.h" />
     <ClInclude Include="..\..\server\SaveLoad.h" />
     <ClInclude Include="..\..\server\ServerApp.h" />
+    <ClInclude Include="..\..\server\ServerFramework.h" />
     <ClInclude Include="..\..\server\ServerFSM.h" />
     <ClInclude Include="..\..\universe\Building.h" />
     <ClInclude Include="..\..\universe\Conditions.h" />
@@ -202,12 +202,12 @@
     <ClCompile Include="..\..\python\EmpireWrapper.cpp" />
     <ClCompile Include="..\..\python\EnumWrapper.cpp" />
     <ClCompile Include="..\..\python\LoggingWrapper.cpp" />
-    <ClCompile Include="..\..\python\server\ServerFramework.cpp" />
     <ClCompile Include="..\..\python\server\ServerWrapper.cpp" />
     <ClCompile Include="..\..\python\UniverseWrapper.cpp" />
     <ClCompile Include="..\..\server\dmain.cpp" />
     <ClCompile Include="..\..\server\SaveLoad.cpp" />
     <ClCompile Include="..\..\server\ServerApp.cpp" />
+    <ClCompile Include="..\..\server\ServerFramework.cpp" />
     <ClCompile Include="..\..\server\ServerFSM.cpp" />
     <ClCompile Include="..\..\universe\UniverseGenerator.cpp" />
     <ClCompile Include="..\..\util\Process.cpp" />
